refactor: 스마트 메뉴 리팩토링
